v.0.6.0.0 Finished separation of pages for CorrectionParameters. Added Page MiscellaneousFunctions, with conversion and database reset. Transfer to SQLite database finished; program now works everywhere with SQLite. Abandoned flat text database. Kept in the codebase but modified to the point that can't work further.

+++ SharedGlucoMan/DataLayer/FlatText/Flat_DataLayerConstructorsAndGeneral.cs
@@ -12,7 +12,7 @@ namespace GlucoMan
         {
             throw new NotImplementedException();
         }
-        internal override int? SaveParameter(string FieldName, string FieldValue, int? Key = null)
+        internal override long? SaveParameter(string FieldName, string FieldValue, int? Key = null)
         {
             throw new NotImplementedException();
         }
